+++ src/main/java/com/johnsunday/app/dao/IPayrollDao.java
@@ -10,13 +10,13 @@ import com.johnsunday.app.entity.Employee;
 import com.johnsunday.app.entity.Payroll;
 
 @Repository
-public interface IPayrollDao extends JpaRepository<Payroll, Integer> {
+public interface IPayrollDao extends JpaRepository<Payroll, Long> {
 
 	// @Query(value="SELECT * FROM payroll WHERE payroll.employee_id_fk=?1",
 	// nativeQuery=true)
 	// List<Payroll> findAllPayrollByEmployeeId(Integer employeeId);
 
-	List<Payroll> findAllByEmployeeId(Integer employeeId);
+	List<Payroll> findAllByEmployeeId(Long employeeId);
 
 	Boolean findByDateAndEmployeeAllIgnoreCase(LocalDateTime payrollDate, Employee employee);
 }
