+++ src/main/java/com/johnsunday/app/security/controller/RoleControllerImpl.java
@@ -20,64 +20,73 @@ import org.springframework.web.bind.annotation.RestController;
 import com.johnsunday.app.security.entity.Role;
 import com.johnsunday.app.security.service.RoleServiceImpl;
 
-@CrossOrigin(origins="*")
+@CrossOrigin(origins = "*")
 @RequestMapping("api/v1/role")
 @RestController
-public class RoleControllerImpl implements IRoleController {	
-	
+public class RoleControllerImpl implements IRoleController {
+
 	@Lazy
 	@Autowired
 	private RoleServiceImpl roleService;
-	
+
 	@PreAuthorize("hasRole('ROLE_ADMIN')")
-	@GetMapping("/")	
+	@GetMapping("/")
 	public ResponseEntity<?> getAllRole() {
 		try {
 			return ResponseEntity.status(HttpStatus.OK).body(roleService.findAll());
-		}catch(Exception e) {
+		} catch (Exception e) {
 			e.printStackTrace();
-			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("{\"error\":\"Error. Please, Try it later. It is NOT possible to SHOW all the roles\"}");
+			return ResponseEntity.status(HttpStatus.NOT_FOUND)
+					.body("{\"error\":\"Error. Please, Try it later. It is NOT possible to SHOW all the roles\"}");
 		}
 	}
+
 	@PreAuthorize("hasAnyRole('ROLE_ADMIN')")
 	@GetMapping("/{roleId}")
-	public ResponseEntity<?> getRoleById(@PathVariable("roleId")Integer roleId) {
+	public ResponseEntity<?> getRoleById(@PathVariable("roleId") Long roleId) {
 		try {
 			return ResponseEntity.status(HttpStatus.OK).body(roleService.findById(roleId));
-		}catch(Exception e) {
+		} catch (Exception e) {
 			e.printStackTrace();
-			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("{\"error\":\"Error. Please, Try it later. It is NOT possible to SHOW the role who you find.\"}");
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).body(
+					"{\"error\":\"Error. Please, Try it later. It is NOT possible to SHOW the role who you find.\"}");
 		}
 	}
+
 	@PreAuthorize("hasAnyRole('ROLE_ADMIN')")
-	@PostMapping("/")	
+	@PostMapping("/")
 	public ResponseEntity<?> saveRole(@Valid @RequestBody Role role) {
 		try {
 			return ResponseEntity.status(HttpStatus.OK).body(roleService.save(role));
-		}catch(Exception e) {
+		} catch (Exception e) {
 			e.printStackTrace();
-			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("{\"error\":\"Error. Please, Try it later. NOT possible to SAVE the user.\"}");
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+					.body("{\"error\":\"Error. Please, Try it later. NOT possible to SAVE the user.\"}");
 		}
 	}
+
 	@PreAuthorize("hasRole('ROLE_ADMIN')")
-	@PutMapping("/{roleId}")	
+	@PutMapping("/{roleId}")
 	public ResponseEntity<?> updateRole(@RequestBody Role role,
-										@PathVariable("roleId")Integer roleId) {
+			@PathVariable("roleId") Long roleId) {
 		try {
-			return ResponseEntity.status(HttpStatus.OK).body(roleService.update(roleId,role));
-		}catch(Exception e) {
+			return ResponseEntity.status(HttpStatus.OK).body(roleService.update(roleId, role));
+		} catch (Exception e) {
 			e.printStackTrace();
-			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("{\"error\":\"Error. Please, Try it later. it is NOT possible UPDATE the user who you looking for.\"}");
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(
+					"{\"error\":\"Error. Please, Try it later. it is NOT possible UPDATE the user who you looking for.\"}");
 		}
 	}
+
 	@PreAuthorize("hasRole('ROLE_ADMIN')")
-	@DeleteMapping("/{roleId}")	
-	public ResponseEntity<?> deleteRole(@PathVariable("roleId")Integer roleId) {
+	@DeleteMapping("/{roleId}")
+	public ResponseEntity<?> deleteRole(@PathVariable("roleId") Long roleId) {
 		try {
 			return ResponseEntity.status(HttpStatus.NO_CONTENT).body(roleService.delete(roleId));
-		}catch(Exception e) {
+		} catch (Exception e) {
 			e.printStackTrace();
-			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("{\"error\":\"Error. Please, Try it later. It is NOT possible DELETE the user who you looking for.\"}");
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(
+					"{\"error\":\"Error. Please, Try it later. It is NOT possible DELETE the user who you looking for.\"}");
 		}
 	}
 }
